tests/requirements.txt
nose
mockredispy